# openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/project": {
      "post": {
        "operationId": "ProjectController_create",
        "summary": "Create a project",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The project has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          }
        },
        "tags": [
          "Project"
        ]
      },
      "get": {
        "operationId": "ProjectController_findAll",
        "summary": "Get all projects with pagination and search.",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "company",
            "required": false,
            "in": "query",
            "description": "Include company",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "university",
            "required": false,
            "in": "query",
            "description": "Include university",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Search by name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "required": false,
            "in": "query",
            "description": "Search by description",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectPreviewDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Project"
        ]
      }
    },
    "/project/{id}": {
      "get": {
        "operationId": "ProjectController_findOne",
        "summary": "Get a project by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "404": {
            "description": "The record does not exist"
          }
        },
        "tags": [
          "Project"
        ]
      },
      "patch": {
        "operationId": "ProjectController_update",
        "summary": "Update a project",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The project has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                }
              }
            }
          },
          "404": {
            "description": "The record does not exist"
          }
        },
        "tags": [
          "Project"
        ]
      },
      "delete": {
        "operationId": "ProjectController_remove",
        "summary": "Delete Project",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The record has been successfully deleted."
          },
          "404": {
            "description": "The record does not exist"
          }
        },
        "tags": [
          "Project"
        ]
      }
    },
...
    "/professor": {
      "post": {
        "operationId": "ProfessorController_create",
        "summary": "Create a professor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfessorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The professor has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessorPreviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "Professor"
        ]
      },
      "get": {
        "operationId": "ProfessorController_findAll",
        "summary": "Get all professors with pagination and search",
        "parameters": [
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Number of records to take",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Search by name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "required": false,
            "in": "query",
            "description": "Search by description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "students",
            "required": false,
            "in": "query",
            "description": "Include students data in the result",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "university",
            "required": false,
            "in": "query",
            "description": "Include university data in the result",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The professors have been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProfessorPreviewDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Professor"
        ]
      }
    },
    "/professor/{id}": {
      "get": {
        "operationId": "ProfessorController_findOne",
        "summary": "Get a professor by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The professor has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Professor"
        ]
      },
      "patch": {
        "operationId": "ProfessorController_update",
        "summary": "Update a professor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfessorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The professor has been successfully updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessorDto"
                }
              }
            }
          }
        },
        "tags": [
          "Professor"
        ]
      },
      "delete": {
        "operationId": "ProfessorController_remove",
        "summary": "Delete a professor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The professor has been successfully deleted."
          }
        },
        "tags": [
          "Professor"
        ]
      }
    },
...
  },
  "info": {
    "title": "YAssignment",
    "description": "YAssignment's API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Server"
    }
  ],
  "components": {
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "company": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "university": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "ProjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "university": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "id",
              "name",
              "description"
            ],
            "nullable": true
          },
          "company": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "id",
              "name",
              "description"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "university",
          "company"
        ]
      },
      "ProjectPreviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "university": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "id",
              "name",
              "description"
            ],
            "nullable": true
          },
          "company": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "id",
              "name",
              "description"
            ],
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "company": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "university": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          }
        }
      },
...
      "CreateProfessorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "university": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          }
        },
        "required": [
          "name",
          "description",
          "university"
        ]
      },
      "ProfessorPreviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "university": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "id",
              "name",
              "description"
            ],
            "nullable": true
          },
          "students": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "minLength": 1
                },
                "name": {
                  "type": "string",
                  "minLength": 1
                },
                "description": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "required": [
                "id",
                "name",
                "description"
              ]
            }
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ]
      },
      "ProfessorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "university": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "minLength": 1
              },
              "name": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              }
            },
            "required": [
              "id",
              "name",
              "description"
            ],
            "nullable": true
          },
          "students": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "minLength": 1
                },
                "name": {
                  "type": "string",
                  "minLength": 1
                },
                "description": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "required": [
                "id",
                "name",
                "description"
              ]
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "university",
          "students"
        ]
      },
      "UpdateProfessorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "university": {
            "type": "string",
            "format": "uuid",
            "minLength": 1
          }
        }
      },
...
    }
  }
}